refactor: separate storage settings and bound refresh lifecycles
