+++ src/components/setting/__tests__/StorageSection.test.tsx
@@ -4,7 +4,8 @@ import userEvent from '@testing-library/user-event'
 import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest'
 import { getSearchStatus } from '@/api/daemon'
 import * as storageApi from '@/api/storage'
-import StorageSection, { resolveMaxItemsCount } from '@/components/setting/StorageSection'
+import { resolveMaxItemsCount } from '@/components/setting/retention-policy'
+import StorageSection from '@/components/setting/StorageSection'
 import { useSetting } from '@/hooks/useSetting'
 import i18n from '@/i18n'
 import { makeBaseSettings } from '@/test/fixtures/settings'
